requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==1.10.12
python-dotenv==1.0.0
bcrypt==4.1.0
PyJWT==2.8.0
